chore: hook機能関連コードをhookServiceクラスに独立させた
このまま追記し続けるとソースコードが長くなりそうだったため
